server-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.12
langgraph==0.0.15
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2